31aug2023

+++ server/seed.py
@@ -7,7 +7,7 @@ def clear_files(directory_path):
     for filename in os.listdir(directory_path):
         file_path = os.path.join(directory_path,filename)
         try:
-            if os.path.isfile(file_path):
+            if os.path.isfile(file_path) and filename != '.gitkeep':
                 os.unlink(file_path)
         except Exception as e:
             print(f'Error deleting {file_path}: {e}')
